Start using the factorized functions through Hydra
